update dependencies and UI

+++ tests/UI/crud_trello_board.test.ts
@@ -2,7 +2,11 @@ import { expect } from '@playwright/test'
 import { test } from '../../support-code/pages/index'
 import TestDataGenerator from '../../support-code/utilities/test_data_generator'
 
-test.beforeEach(async ({trello}) => {
+const boardName = TestDataGenerator.generateBoardName()
+const updatedBoardName = TestDataGenerator.generateBoardName()
+let boardDeleted = false
+
+test.beforeEach(async ({ trello }) => {
     await trello.homePage.goto()
     await (await trello.header.getLoginBtn()).click()
     await trello.loginPage.login(
@@ -12,10 +16,7 @@ test.beforeEach(async ({trello}) => {
     await expect(await trello.homePage.getSectionHeader()).toContainText('YOUR WORKSPACES', { timeout: 30000 })
 })
 
-test('CRUD Trello board through UI', async ({trello}) => {
-    const boardName = TestDataGenerator.generateBoardName()
-    const updatedBoardName = TestDataGenerator.generateBoardName()
-
+test('CRUD Trello board through UI', async ({ trello }) => {
     await test.step('create board', async () => {
         await trello.homePage.createNewBoard(boardName, '🌈')
         await trello.boardPage.waitForPageLoaded()
@@ -33,23 +34,28 @@ test('CRUD Trello board through UI', async ({trello}) => {
     await test.step('close board', async () => {
         await trello.workSpaceNav.closeCurrentBoard()
 
-        await expect(await trello.boardPage.getCloseBoardMessage()).toContainText(`${updatedBoardName} is closed.`)
+        await expect(await trello.boardPage.getCloseBoardMessage()).toContainText('This board is closed. Reopen the board to make changes.')
     })
 
     await test.step('delete board', async () => {
         await trello.boardPage.deleteBoard()
+        await (await trello.header.getTrelloBtn()).click()
 
         await expect(await trello.homePage.getSectionHeader()).toContainText('YOUR WORKSPACES')
-        await expect(await trello.homePage.getAllBoardNames()).not.toContain(updatedBoardName)
+        expect(await trello.homePage.getAllBoardNames()).not.toContain(updatedBoardName)
+
+        boardDeleted = true
     })
 })
 
-test.afterAll(async ({API}) => {
+test.afterAll(async ({ API }) => {
     const token = `${process.env.API_TOKEN}`
     const key = `${process.env.API_KEY}`
     const boards = await API.members.getBoardsFromMember(key, token)
     for (const board of boards) {
+        if (board.name == boardName && boardDeleted == false) {
             const responseStatus = await API.boards.deleteBoard(board.id, key, token)
             expect(responseStatus).toBe(200)
+        }
     }
 })
